Renames ViewModels
