+++ src/Iot.WebApi/Mappings/Devices/DeviceMultipleSensorsMappingProfile.cs
@@ -2,7 +2,6 @@
 using Iot.Domain.Interfaces;
 using Iot.Domain.Models;
 using Iot.WebApi.Responses;
-using Iot.WebApi.ViewModels;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -12,16 +11,16 @@ namespace Iot.WebApi.Mappings.Devices
     {
         public static DeviceMultipleSensorsResponse Map(string deviceId, IEnumerable<IMeasurement> measurements)
         {
-            var sensorsVM = new List<SensorViewModel>();
+            var sensorsVM = new List<SensorResponse>();
 
             foreach (SensorType sensorType in measurements.Select(m => m.SensorType).Distinct())
             {
-                var measurementsVM = new List<MeasurementViewModel>();
+                var measurementsVM = new List<MeasurementResponse>();
                 foreach (var measurement in measurements.Where(m => m.SensorType == sensorType))
                 {
-                    measurementsVM.Add(new MeasurementViewModel() { Date = measurement.FullDate, Value = measurement.Value });
+                    measurementsVM.Add(new MeasurementResponse() { Date = measurement.FullDate, Value = measurement.Value });
                 }
-                sensorsVM.Add(new SensorViewModel() { SensorType = sensorType, Measurements = measurementsVM });
+                sensorsVM.Add(new SensorResponse() { SensorType = sensorType, Measurements = measurementsVM });
             }
 
             var response = new DeviceMultipleSensorsResponse()
