+++ src/Iot.WebApi/Responses/DeviceMultipleSensorsResponse.cs
@@ -1,11 +1,10 @@
-﻿using Iot.WebApi.ViewModels;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace Iot.WebApi.Responses
 {
     public class DeviceMultipleSensorsResponse : BaseResponse
     {
         public string Device { get; set; }
-        public IEnumerable<SensorViewModel> Sensors { get; set; }
+        public IEnumerable<SensorResponse> Sensors { get; set; }
     }
 }
