+++ src/Iot.WebApi/Responses/DeviceSingleSensorResponse.cs
@@ -1,10 +1,8 @@
-﻿using Iot.WebApi.ViewModels;
-
-namespace Iot.WebApi.Responses
+﻿namespace Iot.WebApi.Responses
 {
     public class DeviceSingleSensorResponse : BaseResponse
     {
         public string Device { get; set; }
-        public SensorViewModel Sensor { get; set; }
+        public SensorResponse Sensor { get; set; }
     }
 }
